6. Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -11,11 +11,11 @@ int main(void)
 	long int l;
 	long long int ll;
 	float f;
-	printf("Size of a char: %lu byte(s)\n", sizeof(c));
-	printf("Size of an int: %lu byte(s)\n", sizeof(b));
-	printf("Size of a long int: %lu byte(s)\n", sizeof(l));
-	printf("Size of a long Long int: %lu byte(s)\n", sizeof(ll));
-	printf("Size of a float: %lu byte(s)\n", sizeof(f));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of an int: %lu byte(s)\n",(unsigned long)sizeof(b));
+	printf("Size of a long int: %lu byte(s)\n",(unsigned long)sizeof(l));
+	printf("Size of a long Long int: %lu byte(s)\n",(unsigned long)sizeof(ll));
+	printf("Size of a float: %lu byte(s)\n",(unsigned long)sizeof(f));
 	return (0);
 }
